+++ inc/libalx/base/math/factorial.h
@@ -46,7 +46,7 @@ long double	alx_ldbl_factorial		(int16_t n);
 double		alx_factorial			(int16_t n);
 float		alx_flt_factorial		(int16_t n);
 int		alx_factorial_factorized	(int16_t n,
-				int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
+				int16_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
 
 
 /******************************************************************************
