+++ inc/libalx/base/math/prime_defactorization.h
@@ -55,27 +55,27 @@ static inline	long double	alx_ldbl_prime_defactorization_s8(
 static inline	long double	alx_ldbl_prime_defactorization_u8(
 			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_U8]);
 static inline	long double	alx_ldbl_prime_defactorization_s16(
-			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
+			const int16_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
 static inline	long double	alx_ldbl_prime_defactorization_s16(
-			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
+			const int16_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
 
 static inline	double		alx_prime_defactorization_s8(
 			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S8]);
 static inline	double		alx_prime_defactorization_u8(
 			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_U8]);
 static inline	double		alx_prime_defactorization_s16(
-			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
+			const int16_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
 static inline	double		alx_prime_defactorization_s16(
-			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
+			const int16_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
 
 static inline	float		alx_flt_prime_defactorization_s8(
 			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S8]);
 static inline	float		alx_flt_prime_defactorization_u8(
 			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_U8]);
 static inline	float		alx_flt_prime_defactorization_s16(
-			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
+			const int16_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
 static inline	float		alx_flt_prime_defactorization_s16(
-			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
+			const int16_t (*restrict pf)[PRIME_NUMBERS_QTY_S16]);
 
 
 /******************************************************************************
@@ -109,7 +109,7 @@ long double	alx_ldbl_prime_defactorization_u8(
 
 static inline
 long double	alx_ldbl_prime_defactorization_s16(
-			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S16])
+			const int16_t (*restrict pf)[PRIME_NUMBERS_QTY_S16])
 {
 	long double	n;
 
@@ -122,7 +122,7 @@ long double	alx_ldbl_prime_defactorization_s16(
 
 static inline
 long double	alx_ldbl_prime_defactorization_u16(
-			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_U16])
+			const int16_t (*restrict pf)[PRIME_NUMBERS_QTY_U16])
 {
 	long double	n;
 
@@ -162,7 +162,7 @@ double		alx_prime_defactorization_u8(
 
 static inline
 double		alx_prime_defactorization_s16(
-			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S16])
+			const int16_t (*restrict pf)[PRIME_NUMBERS_QTY_S16])
 {
 	double_t	n;
 
@@ -175,7 +175,7 @@ double		alx_prime_defactorization_s16(
 
 static inline
 double		alx_prime_defactorization_u16(
-			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_U16])
+			const int16_t (*restrict pf)[PRIME_NUMBERS_QTY_U16])
 {
 	double_t	n;
 
@@ -215,7 +215,7 @@ float		alx_flt_prime_defactorization_u8(
 
 static inline
 float		alx_flt_prime_defactorization_s16(
-			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S16])
+			const int16_t (*restrict pf)[PRIME_NUMBERS_QTY_S16])
 {
 	float_t	n;
 
@@ -228,7 +228,7 @@ float		alx_flt_prime_defactorization_s16(
 
 static inline
 float		alx_flt_prime_defactorization_u16(
-			const int8_t (*restrict pf)[PRIME_NUMBERS_QTY_U16])
+			const int16_t (*restrict pf)[PRIME_NUMBERS_QTY_U16])
 {
 	float_t	n;
