+++ src/base/math/prime_factorization.c
@@ -128,7 +128,7 @@ int	alx_prime_factorization_u8	(uint8_t n,
 }
 
 int	alx_prime_factorization_s16	(int16_t n,
-				int8_t (*restrict pf)[PRIME_NUMBERS_QTY_S16])
+				int16_t (*restrict pf)[PRIME_NUMBERS_QTY_S16])
 {
 	ptrdiff_t	prime_index;
 	int_fast16_t	tmp;
@@ -173,7 +173,7 @@ int	alx_prime_factorization_s16	(int16_t n,
 }
 
 int	alx_prime_factorization_u16	(uint16_t n,
-				int8_t (*restrict pf)[PRIME_NUMBERS_QTY_U16])
+				int16_t (*restrict pf)[PRIME_NUMBERS_QTY_U16])
 {
 	ptrdiff_t	prime_index;
 	uint_fast16_t	tmp;
